src/frontend/backend/requirements.txt
langchain-0.0.47
flask-2.3.2
flask_cors-4.0.0
pytube-15.0.0
google-3.0.0
openai-0.27.8 
selenium-4.10.0